fix: better code
fix: rm base scrollbar
